Backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
faiss-cpu==1.13.2
numpy==1.26.4
fastembed==0.3.6
python-dotenv==1.0.1